move currencies to data
