+++ src/data/currencies.js
